+++ packages/core/src/parsers/jsx/element-parser.ts
@@ -1,0 +1,230 @@
+import * as babel from '@babel/core';
+import generate from '@babel/generator';
+import { createMitosisNode } from '../../helpers/create-mitosis-node';
+import { JSONOrNode } from '../../types/json';
+import { MitosisNode } from '../../types/mitosis-node';
+
+const { types } = babel;
+
+/**
+ * Parses a JSX element into a MitosisNode.
+ */
+export const jsxElementToJson = (
+  node:
+    | babel.types.JSXElement
+    | babel.types.JSXText
+    | babel.types.JSXFragment
+    | babel.types.JSXExpressionContainer,
+): MitosisNode | null => {
+  if (types.isJSXText(node)) {
+    return createMitosisNode({
+      properties: {
+        _text: node.value,
+      },
+    });
+  }
+  if (types.isJSXExpressionContainer(node)) {
+    if (types.isJSXEmptyExpression(node.expression)) {
+      return null;
+    }
+    // foo.map -> <For each={foo}>...</For>
+    if (
+      types.isCallExpression(node.expression) ||
+      types.isOptionalCallExpression(node.expression)
+    ) {
+      const callback = node.expression.arguments[0];
+      if (types.isArrowFunctionExpression(callback)) {
+        if (types.isIdentifier(callback.params[0])) {
+          const forArguments = callback.params
+            .map((param) => (param as babel.types.Identifier)?.name)
+            .filter(Boolean);
+          return createMitosisNode({
+            name: 'For',
+            bindings: {
+              each: {
+                code: generate(node.expression.callee)
+                  .code // Remove .map or potentially ?.map
+                  .replace(/\??\.map$/, ''),
+              },
+            },
+            scope: {
+              For: forArguments,
+            },
+            properties: {
+              _forName: forArguments[0],
+              _indexName: forArguments[1],
+              _collectionName: forArguments[2],
+            },
+            children: [jsxElementToJson(callback.body as any)!],
+          });
+        }
+      }
+    }
+
+    // {foo && <div />} -> <Show when={foo}>...</Show>
+    if (types.isLogicalExpression(node.expression)) {
+      if (node.expression.operator === '&&') {
+        return createMitosisNode({
+          name: 'Show',
+          bindings: {
+            when: { code: generate(node.expression.left).code! },
+          },
+          children: [jsxElementToJson(node.expression.right as any)!],
+        });
+      } else {
+        // TODO: good warning system for unsupported operators
+      }
+    }
+
+    // {foo ? <div /> : <span />} -> <Show when={foo} else={<span />}>...</Show>
+    if (types.isConditionalExpression(node.expression)) {
+      return createMitosisNode({
+        name: 'Show',
+        meta: {
+          else: jsxElementToJson(node.expression.alternate as any)!,
+        },
+        bindings: {
+          when: { code: generate(node.expression.test).code! },
+        },
+        children: [jsxElementToJson(node.expression.consequent as any)!],
+      });
+    }
+
+    // TODO: support {foo ? bar : baz}
+
+    return createMitosisNode({
+      bindings: {
+        _text: { code: generate(node.expression).code },
+      },
+    });
+  }
+
+  if (types.isJSXFragment(node)) {
+    return createMitosisNode({
+      name: 'Fragment',
+      children: node.children.map((item) => jsxElementToJson(item as any)).filter(Boolean) as any,
+    });
+  }
+
+  const nodeName = generate(node.openingElement.name).code;
+
+  if (nodeName === 'Show') {
+    const whenAttr: babel.types.JSXAttribute | undefined = node.openingElement.attributes.find(
+      (item) => types.isJSXAttribute(item) && item.name.name === 'when',
+    ) as any;
+
+    const elseAttr: babel.types.JSXAttribute | undefined = node.openingElement.attributes.find(
+      (item) => types.isJSXAttribute(item) && item.name.name === 'else',
+    ) as any;
+
+    const whenValue =
+      whenAttr && types.isJSXExpressionContainer(whenAttr.value)
+        ? generate(whenAttr.value.expression).code
+        : undefined;
+
+    const elseValue =
+      elseAttr &&
+      types.isJSXExpressionContainer(elseAttr.value) &&
+      jsxElementToJson(elseAttr.value.expression as any);
+
+    return createMitosisNode({
+      name: 'Show',
+      meta: {
+        else: elseValue || undefined,
+      },
+      bindings: {
+        ...(whenValue ? { when: { code: whenValue } } : {}),
+      },
+      children: node.children.map((item) => jsxElementToJson(item as any)).filter(Boolean) as any,
+    });
+  }
+
+  // <For ...> control flow component
+  if (nodeName === 'For') {
+    const child = node.children.find((item) => types.isJSXExpressionContainer(item));
+    if (types.isJSXExpressionContainer(child)) {
+      const childExpression = child.expression;
+
+      if (types.isArrowFunctionExpression(childExpression)) {
+        const forArguments = childExpression?.params
+          .map((param) => (param as babel.types.Identifier)?.name)
+          .filter(Boolean);
+
+        return createMitosisNode({
+          name: 'For',
+          bindings: {
+            each: {
+              code: generate(
+                (
+                  (node.openingElement.attributes[0] as babel.types.JSXAttribute)
+                    .value as babel.types.JSXExpressionContainer
+                ).expression,
+              ).code,
+            },
+          },
+          scope: {
+            For: forArguments,
+          },
+          properties: {
+            _forName: forArguments[0],
+            _indexName: forArguments[1],
+            _collectionName: forArguments[2],
+          },
+          children: [jsxElementToJson(childExpression.body as any)!],
+        });
+      }
+    }
+  }
+
+  return createMitosisNode({
+    name: nodeName,
+    properties: node.openingElement.attributes.reduce((memo, item) => {
+      if (types.isJSXAttribute(item)) {
+        const key = item.name.name as string;
+        const value = item.value;
+        if (types.isStringLiteral(value)) {
+          memo[key] = value;
+          return memo;
+        }
+        if (types.isJSXExpressionContainer(value) && types.isStringLiteral(value.expression)) {
+          memo[key] = value.expression.value;
+          return memo;
+        }
+      }
+      return memo;
+    }, {} as { [key: string]: JSONOrNode }) as any,
+    bindings: node.openingElement.attributes.reduce((memo, item) => {
+      if (types.isJSXAttribute(item)) {
+        const key = item.name.name as string;
+        const value = item.value;
+
+        if (types.isJSXExpressionContainer(value) && !types.isStringLiteral(value.expression)) {
+          const { expression } = value;
+          if (types.isArrowFunctionExpression(expression)) {
+            if (key.startsWith('on')) {
+              memo[key] = {
+                code: generate(expression.body).code,
+                arguments: expression.params.map((node) => (node as babel.types.Identifier)?.name),
+              };
+            } else {
+              memo[key] = { code: generate(expression.body).code };
+            }
+          } else {
+            memo[key] = { code: generate(expression).code };
+          }
+
+          return memo;
+        }
+      } else if (types.isJSXSpreadAttribute(item)) {
+        // TODO: potentially like Vue store bindings and properties as array of key value pairs
+        // too so can do this accurately when order matters. Also tempting to not support spread,
+        // as some frameworks do not support it (e.g. Angular) tho Angular may be the only one
+        memo._spread = {
+          code: types.stringLiteral(generate(item.argument).code),
+        };
+      }
+      return memo;
+    }, {} as { [key: string]: JSONOrNode }) as any,
+    children: node.children.map((item) => jsxElementToJson(item as any)).filter(Boolean) as any,
+  });
+};
